Analyzing the file path

+++ php/action_upload.php
@@ -55,22 +55,44 @@ $up = new Uploader($fieldName, $config, $base64);
 $info = $up->getFileInfo();
 
 // 将文件同步存储到 UPYUN
+/*
+function log($log) {
+    $log_file = "log.txt";
+
+    $handle = fopen($log_file, "a");
+    fwrite($handle, $log);
+    fclose($handle);
+    exit;
+}
+ */
+
 $upyun = new UpYun($bucketname, $username, $password);
 
 try {
     $uri = strstr($info["url"], "upload");
 
-    $opts = array(
-        UpYun::CONTENT_MD5 => md5(file_get_contents($uri))
-    );
+    if (file_exists($uri)) {
+        $opts = array(
+            UpYun::CONTENT_MD5 => md5(file_get_contents($uri))
+        );
 
-    $fh = fopen($uri, "rb");
-    $rsp = $upyun->writeFile("/ueditor/php/" . $uri, $fh, True, $opts);
-    fclose($fh);
+        $fh = fopen($uri, "rb");
+        $rsp = $upyun->writeFile("/ueditor/php/" . $uri, $fh, True, $opts);
+        fclose($fh);
+    }
+    else {
+        $log = date("Y-m-d H:m:s") . " 文件不存在，请检查目录是否正确。" . "\r\n";
+        $log_file = "log.txt";
+
+        $handle = fopen($log_file, "a");
+        fwrite($handle, $log);
+        fclose($handle);
+        exit;
+    }
 }
 catch(Exception $e) {
-    $log_file = "log.txt";
     $log = date("Y-m-d H:m:s") . " " . $e->getCode() . " " . $e->getMessage() . "\r\n";
+    $log_file = "log.txt";
 
     $handle = fopen($log_file, "a");
     fwrite($handle, $log);
